requirements.txt
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
cryptography==41.0.7
cvzone==1.6.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.2
dlib==19.24.2
executing==2.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.105.0
fastjsonschema==2.19.0
filelock==3.13.1
firebase-admin==6.3.0
Flask==3.0.0
Flask-Jsonpify==1.5.0
Flask-Login==0.6.3
Flask-MySQL==1.5.2
fqdn==1.5.1
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.3
httplib2==0.22.0
httpx==0.26.0
idna==3.6
imageio==2.33.1
imutils==0.5.4
ipykernel==6.27.1
ipython==8.18.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonify==0.5
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
login==0.0.6
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
MouseInfo==0.1.3
msgpack==1.0.7
mysql-connector-python==8.3.0
nbclient==0.9.0
nbconvert==7.13.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.2
opencv-python==4.8.1.78
overrides==7.4.0
packaging==23.2
pandocfilters==1.5.0
parso==0.8.3
Pillow==10.1.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.7
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyGetWindow==0.0.9
Pygments==2.17.2
PyJWT==2.8.0
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyngrok==7.0.3
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pyserial==3.5
python-dateutil==2.8.2
python-firebase==1.2
python-json-logger==2.0.7
python-telegram-bot==20.8
pytweening==1.0.7
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.15.2
rsa==4.9
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.27.0
terminado==0.18.0
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.25.0
virtualenv==20.25.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
